change how queries and templates are linked to views
Queries and templates now has to be explicitly stated in the view's yaml. This to make it possible to reuse queries/templates, in the future to make it possible to support template formats other than HTML and partial updates.

Layouts/includes goes in templates/includes, templates goes in templates/*, queries goes in queries/*.