migrate to Cobra
snowman now uses the Cobra library for CLI related features. Building a Snowman site now requires the "build" subcommand. There is now also a --help flag for both the root command and subcommands.

This change also restructures the structure of the project as well as moves it to the glacier-in-archives namespace.