dynamic queries and changes to template input
This introduces a new template function "query" this allows you to issue SPARQL during rendering.

query "location_of_query" argument

A side effect of this is a major change to the way a SPARQL resultset is represented in the templates. There is no longer a need for .Value but in some cases one might need to call .String.