partial/includes templates do no longer need to be in a dedicated folder
includes/partial templates does no longer need to be under views/includes, now all templates are parsed during view rendering no matter if they are standard ones or partial ones. However this causes rendering to be much slower as all templates are parsed on each view render, in the future all templates should only be parsed once and the this will no longer be an issue.