refactor cache implementation
This moves the cache implementation away form the SPARQL client. It will now also use both the query file location and its content. This will allow to invalidate cache for all responses for a particular dynamic query as the location hash wont depend on the input value. This will also allow for  an user facing interface meant for invalidating cache outside of the build process.

Side effect: the --cached flag has been replaced with a cache flag.