new command to inspect cache
this also removes the clearcache command as its functionallity os now under the new command and the flag invalidate cache

this also enables the "never" cache strategy.

Later down the road it might be of interest to move some of the functionallity from the cache cmd layer down to the cache manager.