Implemeent the server command Create a built in HTTP server `snowman server` #4
This is a very early poc which needs to be refectored.

First the server needs to run in a seperate channel so that we can watch the filesytem and run the livereload server in parallel(#4). Then we need to listen to signals like CTRL+C to gracefully shut the server down.

https://rafallorenz.com/go/handle-signals-to-graceful-shutdown-http-server/