🔧 move template functions to its own package
This moves template functions and their implementations from the utils and view packages. This aims to make the views package simpler and make the functions easier to test.

The functions dependent on repo/site config haven't been moved and to do so we would need to be able to make repo/site config global.

It would be nice if we could merge funcmaps instead of chaining them but I don't have a solution for that at this point.